requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/base.in
#
anyascii==0.1.7           # via wagtail
argon2-cffi==20.1.0       # via -r requirements/base.in
asgiref==3.3.1            # via django
beautifulsoup4==4.8.2     # via wagtail
certifi==2020.11.8        # via elasticsearch, requests
cffi==1.14.3              # via argon2-cffi
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via python3-openid
django-allauth==0.42.0    # via -r requirements/base.in
django-appconf==1.0.4     # via django-compressor
django-compressor==2.4    # via -r requirements/base.in
django-crispy-forms==1.9.2  # via -r requirements/base.in
django-elasticsearch-dsl==7.1.4  # via -r requirements/base.in
django-environ==0.4.5     # via -r requirements/base.in
django-filter==2.4.0      # via wagtail
django-kdl-wagtail==0.9.1  # via -r requirements/base.in
django-model-utils==4.0.0  # via -r requirements/base.in
django-modelcluster==5.1  # via wagtail
django-redis==4.12.1      # via -r requirements/base.in
django-taggit==1.3.0      # via wagtail
django-treebeard==4.3.1   # via wagtail
django==3.0.10            # via -r requirements/base.in, django-allauth, django-appconf, django-filter, django-model-utils, django-redis, django-taggit, django-treebeard, djangorestframework, wagtail
djangorestframework==3.12.2  # via wagtail
draftjs-exporter==2.1.7   # via wagtail
elasticsearch-dsl==7.3.0  # via -r requirements/base.in, django-elasticsearch-dsl
elasticsearch==7.10.0     # via elasticsearch-dsl
et-xmlfile==1.0.1         # via openpyxl
hiredis==1.1.0            # via -r requirements/base.in
html5lib==1.1             # via wagtail
idna==2.10                # via requests
jdcal==1.4.1              # via openpyxl
l18n==2020.6.1            # via wagtail
lxml==4.6.1               # via -r requirements/base.in
oauthlib==3.1.0           # via requests-oauthlib
openpyxl==3.0.5           # via tablib
pillow==7.2.0             # via -r requirements/base.in, wagtail
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via elasticsearch-dsl
python-slugify==4.0.1     # via -r requirements/base.in
python3-openid==3.2.0     # via django-allauth
pytz==2020.1              # via -r requirements/base.in, django, django-modelcluster, l18n
rcssmin==1.0.6            # via -r requirements/base.in, django-compressor
redis==3.5.3              # via -r requirements/base.in, django-redis
regex==2020.11.13         # via -r requirements/base.in
requests-oauthlib==1.3.0  # via django-allauth
requests==2.25.0          # via django-allauth, requests-oauthlib, wagtail
rjsmin==1.1.0             # via django-compressor
six==1.15.0               # via argon2-cffi, django-compressor, django-elasticsearch-dsl, elasticsearch-dsl, html5lib, l18n, python-dateutil
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.4.1           # via django
tablib[xls,xlsx]==2.0.0   # via wagtail
text-unidecode==1.3       # via python-slugify
tqdm==4.51.0              # via -r requirements/base.in
unidecode==1.1.1          # via wagtail
urllib3==1.26.2           # via elasticsearch, requests
wagtail==2.11.1           # via -r requirements/base.in, django-kdl-wagtail
webencodings==0.5.1       # via html5lib
whitenoise==5.2.0         # via -r requirements/base.in
willow==1.4               # via wagtail
xlrd==1.2.0               # via tablib
xlsxwriter==1.3.7         # via wagtail
xlwt==1.3.0               # via tablib
